clase GestionBanco borrada
ya no se necesita
